Implement a dedicated exitfication pass #14152
The idea is described in #14152, and can be summarized: Float the exit
path out of a joinrec, so that the simplifier can do more with it.
See the test case for a nice example.

The floating goes against what the simplifier usually does, hence we
need to be careful not inline them back.

The position of exitification in the pipeline was chosen after a small
amount of experimentation, but may need to be improved. For example,
exitification can allow rewrite rules to fire, but for that it would
have to happen before the `simpl_phases`.

Together with the next patch, perf.haskell.org reports these nice
performance wins:

	Nofib allocations    prev         change       gain
	fannkuch-redux	 78446688	- 99.92%      64608
	k-nucleotide	109466432	- 91.32%    9502064
	simple		 72424696	-  5.96%   68109560

	Nofib instruction counts
	compress2       573354476	+  3.34%  592476104
	k-nucleotide   2310080537	-  5.59% 2180917263
	scs	       1979135192	-   3.2% 1915880589
	simple	        670348375	-   4.9%  637507288

Differential Revision: https://phabricator.haskell.org/D3903