simplNonRecJoinPoint: Handle Shadowing correctly
Previously, (since 33452df), simplNonRecJoinPoint would do the wrong
thing in the presence of shadowing: It analyzed the RHS of a join
binding with the environment for the body. In particular, with

    foo x =
      join x = x * x
      in x

where there is shadowing, it renames the inner x to x1, and should
produce

    foo x =
      join x1 = x * x
      in x1

but because the substitution (x ↦ x1) is also used on the RHS we get the
bogus

    foo x =
      join x1 = x1 * x1
      in x1

The corresponding function for non-join-points, i.e. simplNonRecE, does
the right thing and has done so forever, so I’ll skip creating a test
case or a note for this.

Differential Revision: https://phabricator.haskell.org/D4130